Fix security risk for null emails
The returned email from Facebook is sometimes null. Bagisto tries to match the returned email from Facebook against the database, which would allow logins to accounts that have a null email which may not belong to the Facebook user trying to login, and therefore such user accounts might be breached.

This commit checks for null emails first, and returns a null customer instead of the first customer with a null email.